edit type of variables of quantity from int to float

+++ Kohi/Models/CheckInventoryModel.cs
@@ -11,8 +11,8 @@ namespace Kohi.Models
     {
         public int Id { get; set; }              // Mã phiếu nhập
         public int InventoryId { get; set; }     // Mã lô hàng (liên kết với InventoryModel)
-        public int ActualQuantity { get; set; } // Số lượng thực tế sau khi kiểm kê (không liên quan DB)
-        public int Discrepancy => ActualQuantity - Inventory.Quantity; // Số lượng chênh lệch sau khi kiểm kê (không liên quan DB)
+        public float ActualQuantity { get; set; } // Số lượng thực tế sau khi kiểm kê (không liên quan DB)
+        public float Discrepancy => ActualQuantity - Inventory.Quantity; // Số lượng chênh lệch sau khi kiểm kê (không liên quan DB)
         public DateTime CheckDate { get; set; } // Ngày kiểm kê
         public string? Notes { get; set; }        // Ghi chú (tùy chọn)
