+++ Kohi/Models/InboundModel.cs
@@ -11,7 +11,7 @@ namespace Kohi.Models
     {
         public int Id { get; set; }              // Mã phiếu nhập
         public int IngredientId { get; set; }    // Mã nguyên liệu (liên kết với InventoryModel)
-        public int Quantity { get; set; }        // Số lượng nhập
+        public float Quantity { get; set; }        // Số lượng nhập
         public DateTime InboundDate { get; set; } // Ngày nhập kho
         public DateTime ExpiryDate { get; set; }  // Ngày hết hạn của lô hàng nhập
         public int SupplierId { get; set; }     // Nhà cung cấp (tùy chọn)
